Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
python-dotenv
passlib[bcrypt]
reportlab==4.0.7
